requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration model - gpt-5 mini
# model="gpt-5-mini"
# embedding="text-embedding-3-small"
#
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#pydantic
pydantic
pydantic[email]